Add arg for yaml template from outside mache
The new `yaml_template` argument can point to a yaml file from
a different package that may have compiler or other package
versions that aren't the E3SM defaults.  This is useful if the
E3SM compilers or other packages are too out-of-date to build
certain libraries, as we are finding with Trilinos on some
machines.