Fix handling of "route" key in CLI spec
We add "deploy" as a possible value and fix its handling in
`deploy.py.j2`.

We also add a `--bootstrap-only` flag.

We perform some clean-up and consolidation in `deploy.py.j2`.

We make the download of `bootstrap.py` more robust againt caching
and download only the branch we want to install from to save
(a lot of) time.